backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
PyYAML